Fix some misc issues caused by previous changes
I dont like eclipse

+++ src/main/java/com/blakebr0/extendedcrafting/client/gui/GuiCraftingCore.java
@@ -34,7 +34,7 @@ public class GuiCraftingCore extends GuiContainer {
 
 	private int getProgressBarScaled(int pixels) {
 		int i = this.tile.getProgress();
-		int j = this.tile.getRecipe().getCost();
+		long j = this.tile.getRecipe().getCost();
 		return (int) (j != 0 && i != 0 ? (long) i * pixels / j : 0);
 	}
